Создал контроллер и перенес туда логику по отрисовке задач
