+++ src/main.js
@@ -1,14 +1,10 @@
 import TripInfo from "./components/trip-info";
 import Menu from "./components/menu";
 import Filter from "./components/filter";
-import Sort from "./components/sort";
-import NewEvent from "./components/new-event";
 import TripList from "./components/trip-list";
-import TripDay from "./components/trip-days";
-import TripEvents from "./components/trip-events";
-import NoTripDays from "./components/no-trip-days";
 import {generateTripEvents} from "./mock/event";
-import {render, RenderPosition, replace} from "./utils/render";
+import {render, RenderPosition} from "./utils/render";
+import TripController from "./controllers/TripController";
 
 const TRIP_DAYS_COUNT = 22;
 const tripEvents = generateTripEvents(TRIP_DAYS_COUNT);
@@ -37,61 +33,8 @@ render(siteHeader, new TripInfo(), RenderPosition.BEFOREEND);
 render(siteNavigationMenuHeader, new Menu(), RenderPosition.AFTEREND);
 render(siteNavigationMenu, new Filter(), RenderPosition.BEFOREEND);
 
-const renderTripEvents = (eventsList, container) => {
-  eventsList.items.forEach((tripEvent) => {
-    const currentEvent = new TripEvents(tripEvent);
-    const eventEditComponent = new NewEvent(tripEvent);
-    render(container, currentEvent, RenderPosition.BEFOREEND);
-
-    const replaceEventToEdit = () => {
-      replace(eventEditComponent, currentEvent);
-    };
-
-    const replaceEditToEvent = () => {
-      replace(currentEvent, eventEditComponent);
-    };
-
-    const onEscKeyDown = (evt) => {
-      const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
-
-      if (isEscKey) {
-        replaceEditToEvent();
-        document.removeEventListener(`keydown`, onEscKeyDown);
-      }
-    };
-
-    currentEvent.setEditButtonClickHandler(() => {
-      replaceEventToEdit();
-      document.addEventListener(`keydown`, onEscKeyDown);
-    });
-
-    eventEditComponent.setSubmitHandler((evt) => {
-      evt.preventDefault();
-      replaceEditToEvent();
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    });
-  });
-};
-
-const renderTripDays = (container, eventDays) => {
-  render(siteTripEvents, new Sort(), RenderPosition.AFTERBEGIN);
-
-  eventDays.forEach((items, index) => {
-    const tripDay = new TripDay(items, index);
-    render(container, tripDay, RenderPosition.BEFOREEND);
-    const tripEventsList = tripDay.getElement().querySelector(`.trip-events__list`);
-    renderTripEvents(items, tripEventsList);
-  });
-};
-
-const renderTripList = (container) => {
-  if (groupEventItems.length > 0) {
-    renderTripDays(container, groupEventItems);
-  } else {
-    render(siteTripEvents, new NoTripDays(), RenderPosition.BEFOREEND);
-  }
-};
-
 const tripList = new TripList();
+const tripController = new TripController(tripList);
+
 render(siteTripEvents, tripList, RenderPosition.BEFOREEND);
-renderTripList(tripList.getElement());
+tripController.render(groupEventItems, siteTripEvents);
